move structure export functions from `lib/io/export.ts` to `lib/structure/export.ts` for better organization
- update import paths in related files to reflect the new structure
