+++ tests/vitest/structure/Structure.test.ts
@@ -1,8 +1,8 @@
 import type { AnyStructure, Vec3 } from '$lib'
 import { Structure } from '$lib'
-import * as exports from '$lib/io/export'
 import { euclidean_dist, type Matrix3x3, pbc_dist } from '$lib/math'
 import { DEFAULTS } from '$lib/settings'
+import * as exports from '$lib/structure/export'
 import { structures } from '$site/structures'
 import { readFileSync } from 'fs'
 import { mount, tick } from 'svelte'
